requirements.txt
blinker==1.6.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.6
colorama==0.4.6
dnspython==2.4.2
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
py.validator==0.19.1
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
timestring==1.6.4
typing==3.7.4.3
urllib3==2.1.0
Werkzeug==2.3.6
zipp==3.16.2
